--- src/session5/Human.java
@@ -1,4 +1,0 @@
-package session5;
-
-public class Human {
-}
