+++ src/session5/homework5/JumpStatements.java
@@ -1,4 +1,4 @@
-package homework5;
+package session5.homework5;
 
 import java.util.Scanner;
