+++ src/session5/practice/Human.java
@@ -1,0 +1,4 @@
+package session5.practice;
+
+public class Human {
+}
