+++ src/session5/practice/SwitchStatements.java
@@ -1,4 +1,4 @@
-package session5;
+package session5.practice;
 
 import java.util.Scanner;
